requirements.txt
numpy
pandas
scipy
fastdtw
simplejson
loguru
joblib
pfc-geometry>=0.2.6
flightdata>=0.2.13
